event, notice component responsive css and other component edit css code

+++ src/components/common-component/breadShop_li_component.jsx
@@ -1,33 +1,41 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
+import { createStructuredSelector } from 'reselect';
+import { connect } from 'react-redux';
 
 import { ShopliWrap } from './breadShop_li_style';
 import axios from '../../utils/axios';
 import { errorhandler } from '../../utils/common';
+import { selectCurrentUser } from '../redux/user/user.selectors';
 
-const BreadShopLi = ({ shopList, shopSeverLike, likeTrue, likeFalse, shopId, breadShopId }) => {
+const BreadShopLi = ({ shopList, shopSeverLike, likeTrue, likeFalse, shopId, breadShopId, currentUser, location, history }) => {
   const changeShopHeart = async () => {
-    try {
-      if (shopSeverLike) {
-        const { status } = await axios.delete(`/bread/shop/favorite/${shopId}`);
-        if (status === 200) {
-          likeFalse(shopId);
+    if (!currentUser) {
+      const comeAddress = encodeURIComponent(location.pathname + location.search);
+      history.push(`/signin?moveAddress=${comeAddress}`);
+    } else {
+      try {
+        if (shopSeverLike) {
+          const { status } = await axios.delete(`/bread/shop/favorite/${shopId}`);
+          if (status === 200) {
+            likeFalse(shopId);
+          }
+        } else {
+          const { status: shopStatus } = await axios.post(`/bread/shop/favorite/${shopId}`);
+          if (shopStatus === 200) {
+            likeTrue(shopId);
+          }
         }
-      } else {
-        const { status: shopStatus } = await axios.post(`/bread/shop/favorite/${shopId}`);
-        if (shopStatus === 200) {
-          likeTrue(shopId);
-        }
+      } catch (err) {
+        errorhandler(err);
       }
-    } catch (err) {
-      errorhandler(err);
     }
   };
 
   return (
     <ShopliWrap>
-      <Link to={`/bread-house/detail/${shopId}`} key={`bread_shop_list${shopId}`}>
+      <Link to={`/bread-house/detail/${shopId}`}>
         <div className="image_wrap">
           <img src={shopList.image} alt={`${shopList.title}의 이미지`} />
         </div>
@@ -40,7 +48,7 @@ const BreadShopLi = ({ shopList, shopSeverLike, likeTrue, likeFalse, shopId, bre
       {breadShopId ? null : (
         <img
           src={shopSeverLike ? 'https://s3.ap-northeast-2.amazonaws.com/image.mercuryeunoia.com/images/web/jisu/+common_icon/heart.png' : 'https://s3.ap-northeast-2.amazonaws.com/image.mercuryeunoia.com/images/web/jisu/+common_icon/spaceheart.png'}
-          alt="좋아요하트이미지"
+          alt="하트 이미지"
           className="heart_image"
           aria-hidden="true"
           onClick={changeShopHeart}
@@ -50,20 +58,30 @@ const BreadShopLi = ({ shopList, shopSeverLike, likeTrue, likeFalse, shopId, bre
   );
 };
 
+BreadShopLi.defaultProps = {
+  shopSeverLike: undefined,
+  likeTrue: undefined,
+  likeFalse: undefined,
+  breadShopId: undefined,
+  currentUser: null,
+  history: undefined,
+  location: undefined
+};
+
 BreadShopLi.propTypes = {
   shopList: PropTypes.instanceOf(Object).isRequired,
   shopSeverLike: PropTypes.bool,
   likeTrue: PropTypes.func,
   likeFalse: PropTypes.func,
   shopId: PropTypes.number.isRequired,
-  breadShopId: PropTypes.string
+  breadShopId: PropTypes.string,
+  currentUser: PropTypes.instanceOf(Object),
+  location: PropTypes.instanceOf(Object),
+  history: PropTypes.instanceOf(Object)
 };
 
-BreadShopLi.defaultProps = {
-  shopSeverLike: undefined,
-  likeTrue: undefined,
-  likeFalse: undefined,
-  breadShopId: undefined
-};
+const shopLiStateToProps = createStructuredSelector({
+  currentUser: selectCurrentUser
+});
 
-export default BreadShopLi;
+export default connect(shopLiStateToProps, null)(BreadShopLi);
